Move error codes and schema to proper places
