+++ saleor/graphql/shop/error_codes.py
@@ -1,0 +1,5 @@
+from enum import Enum
+
+
+class OrderSettingsErrorCode(Enum):
+    INVALID = "invalid"
