+++ saleor/graphql/shop/schema.py
@@ -1,10 +1,13 @@
 import graphene
 
+from ...core.permissions import OrderPermissions
+from ..decorators import permission_required
 from ..translations.mutations import ShopSettingsTranslate
 from .mutations import (
     AuthorizationKeyAdd,
     AuthorizationKeyDelete,
     HomepageCollectionUpdate,
+    OrderSettingsUpdate,
     ShopAddressUpdate,
     ShopDomainUpdate,
     ShopFetchTaxRates,
@@ -13,16 +16,23 @@ from .mutations import (
     StaffNotificationRecipientDelete,
     StaffNotificationRecipientUpdate,
 )
-from .types import Shop
+from .types import OrderSettings, Shop
 
 
 class ShopQueries(graphene.ObjectType):
     shop = graphene.Field(
         Shop, description="Return information about the shop.", required=True
     )
+    order_settings = graphene.Field(
+        OrderSettings, description="Order related settings from site settings."
+    )
 
     def resolve_shop(self, _info):
         return Shop()
+
+    @permission_required(OrderPermissions.MANAGE_ORDERS)
+    def resolve_order_settings(self, info, *args, **_kwargs):
+        return info.context.site.settings
 
 
 class ShopMutations(graphene.ObjectType):
@@ -39,3 +49,5 @@ class ShopMutations(graphene.ObjectType):
     shop_fetch_tax_rates = ShopFetchTaxRates.Field()
     shop_settings_translate = ShopSettingsTranslate.Field()
     shop_address_update = ShopAddressUpdate.Field()
+
+    order_settings_update = OrderSettingsUpdate.Field()
