+++ saleor/order/error_codes.py
@@ -27,7 +27,3 @@ class OrderErrorCode(Enum):
     ZERO_QUANTITY = "zero_quantity"
     INSUFFICIENT_STOCK = "insufficient_stock"
     DUPLICATED_INPUT_ITEM = "duplicated_input_item"
-
-
-class OrderSettingsErrorCode(Enum):
-    INVALID = "invalid"
